# src/Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1",
    "title": "Zoobrilka REST API"
  },
  "host": "zoobrilka-alice-skill.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Poem"
    },
    {
      "name": "Record"
    },
    {
      "name": "User"
    }
  ],
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/poem/{id}": {
      "get": {
        "tags": ["Poem"],
        "summary": "Получить стих",
        "parameters": [{ "in": "path", "name": "id", "required": true, "type": "string" }],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "$ref": "#/definitions/Poem"
                }
              }
            }
          }
        }
      }
    },
    "/api/poem/today": {
      "get": {
        "tags": ["Poem"],
        "summary": "Получить стих дня",
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "$ref": "#/definitions/Poem"
                }
              }
            }
          }
        }
      }
    },
    "/api/record": {
      "post": {
        "tags": ["Record"],
        "summary": "Загрузка новой записи",
        "description": "Content-Type: multipart/form-data",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["record", "userId", "poemId"],
              "properties": {
                "record": { "type": "object" },
                "userId": { "type": "string" },
                "poemId": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "$ref": "#/definitions/Record"
                }
              }
            }
          }
        }
      }
    },
    "/api/record/{id}": {
      "get": {
        "tags": ["Record"],
        "summary": "Получить запись",
        "parameters": [{ "in": "path", "name": "id", "required": true, "type": "string" }],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "$ref": "#/definitions/Record"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Record"],
        "summary": "Удалить запись",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          { "in": "body", "name": "userId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/record/{id}/vote": {
      "post": {
        "tags": ["Record"],
        "summary": "Оценить запись",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "type": "object", "required": ["userId", "vote"], "properties": { "userId": { "type": "string" }, "vote": { "type": "string" } } }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/records/{poemId}": {
      "get": {
        "tags": ["Record"],
        "summary": "Возвращает записи стиха по рейтингу",
        "parameters": [
          { "in": "path", "name": "poemId", "required": true, "type": "string" },
          { "in": "query", "name": "offset", "type": "number" }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Record"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/records": {
      "get": {
        "tags": ["Record"],
        "summary": "Возвращает записи стихов по рейтингу",
        "parameters": [{ "in": "query", "name": "offset", "type": "number" }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Record"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{id}/records": {
      "get": {
        "tags": ["User"],
        "summary": "Возвращает топ записей юзера",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          { "in": "query", "name": "poemId", "type": "string" }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Record"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/records": {
      "get": {
        "tags": ["User"],
        "summary": "Возвращает топ юзеров и их записей",
        "parameters": [
          { "in": "query", "name": "poemId", "type": "string" },
          { "in": "query", "name": "offset", "type": "number" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "string" },
                    "records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Record"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/search/": {
      "get": {
        "tags": ["Poem"],
        "summary": "Поиск",
        "parameters": [
          { "in": "query", "name": "title", "type": "string" },
          { "in": "query", "name": "firstName", "type": "string" },
          { "in": "query", "name": "lastName", "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "response": {
                  "$ref": "#/definitions/Poems"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Poem": {
      "required": ["author", "title", "first_line", "text", "id"],
      "properties": {
        "id": { "type": "integer", "uniqueItems": true },
        "author": { "type": "object", "required": ["firstName", "lastName"], "properties": { "firstName": { "type": "string" }, "lastName": { "type": "string" } } },
        "title": { "type": "string" },
        "text": { "type": "string" },
        "first_line": { "type": "string" },
        "tags": { "type": "array", "items": { "type": "string" } }
      }
    },
    "Poems": {
      "type": "array",
      "items": { "$ref": "#/definitions/Poem" },
      "maxItems": 5
    },
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Record": {
      "type": "object",
      "properties": {
        "id": { "type": "string" },
        "url": { "type": "string" },
        "owner": { "type": "string" },
        "poem": { "type": "string" },
        "rating": { "type": "number" },
        "votes": { "type": "object", "properties": { "userId": { "type": "string" } } }
      }
    }
  }
}